server/requirements.txt
flask
flask-cors
pymongo
python-dotenv
marshmallow
groq
pypdf
python-pptx
moviepy
